I figured out what was going on.
Since nobody had made any changes to i386v-nat.c on sourceware since
the repository was started, we were still getting the version from the
vendor branch.  (If you don't understand how the vendor branch works,
treasure your innocence.)  I naively assumed that, to see what changes
had been introduced in version 1.2, one should diff it with 1.1,
as they say in TV Guide, "with hilarious results".

This change simply restores i386v-nat.c to what it used to be before I
started bungling around.  The text of this revision should be
identical to that of revision 1.1.1.3, which was the former head of
the trunk.  (Yes, 1.1.1.3 was the head of the trunk.  As I said,
treasure your innocence.)